Moved and renamed method into JSScrollIntoView in extensions class.
